fix: packaging refactoring
