+++ go-sdk/internal/api/client_test.go
@@ -3,7 +3,7 @@ package api
 import (
 	"fmt"
 	oauthutils "github.com/keptn/go-utils/pkg/common/oauth2"
-	"github.com/keptn/keptn/go-sdk/pkg/sdk/internal/config"
+	"github.com/keptn/keptn/go-sdk/internal/config"
 	"github.com/stretchr/testify/assert"
 	"net/http"
 	"net/http/httptest"
