+++ go-sdk/internal/api/initializer.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	keptnapi "github.com/keptn/go-utils/pkg/api/utils"
 	"github.com/keptn/keptn/cp-common/api"
-	"github.com/keptn/keptn/go-sdk/pkg/sdk/internal/config"
+	"github.com/keptn/keptn/go-sdk/internal/config"
 	"net/http"
 	"net/url"
 	"strings"
